% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_reaction.R
\name{create_reaction}
\alias{create_reaction}
\title{Create reaction for a given reaction of lipid metabolism}
\usage{
create_reaction(
  substrates,
  template = NULL,
  reaction = "fa_to_coa",
  constraints = character(length(substrates)),
  negate = logical(length(substrates))
)
}
\arguments{
\item{substrates}{\code{list}, with entry names equal to substrates}

\item{template}{\code{data.frame}}

\item{reaction}{\code{character(1)}}

\item{constraints}{\code{character}, with length \code{length(substrates)}}

\item{negate}{\code{logical}, with length \code{length(substrates)}}
}
\value{
list of length 2
}
\description{
The function \code{create_reaction} creates from a given template reaction
the specific reaction given a set of substrates.

The function returns a list of the products of the reaction (in the first
entry of the returned list), together with the \code{template} with the 
updated reactions (in the second entry of the returned list).
}
\details{
If \code{template == NULL}, the function will load a template specific
to the \code{reaction} with mininum information on the reaction.
}
\examples{
FA <- c("FA(14:0(12Me))", "FA(16:0(14Me))", "FA(15:1(9Z)(14Me))",        
    "FA(17:0(16Me))", "FA(12:0(11Me))", "FA(13:0(12Me))", "FA(14:0(13Me))",
    "FA(15:0(14Me))", "FA(16:0(15Me))", "FA(12:0)", "FA(14:0)")
    
template <- list(reaction_name = "CoA biosynthesis", 
    reaction_formula = "M_atp + M_coa + M_fatacid <=> M_ppi + M_amp + M_fataccoa",
    reaction_isReversible = "",
    reaction_geneAssociation = "",
    reaction_pathway = "Acyl-CoA synthetase")
 
## run create_reaction (template = NULL)
create_reaction(substrates = list(FA = FA), template = NULL,
    reaction = "fa_to_coa")
    
## run create_reaction (template = template)
create_reaction(substrates = list(FA = FA), template = template,
    reaction = "fa_to_coa")
}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
   and Thomas Naake, \email{thomasnaake@googlemail.com}
}
