% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_via_chebi.R
\name{create_via_chebi}
\alias{create_via_chebi}
\title{create_via_chebi}
\usage{
create_via_chebi()
}
\description{
Creates ...
}
\details{
Uses ...
}
\examples{
rhea_reactions_order <- read.table("inst/rhea_reactions.tsv", sep = "\t", header = TRUE)
rhea_reactions_order <- rhea_reactions_order[c(1, 5, 9), ]
rhea_reactions <- read.table("inst/extdata/rhea-reactions.txt", sep = "\t", header = FALSE, quote = "")

}
