% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_.create_substrates_combination.R
\name{.check_colnames_substrates_combinations}
\alias{.check_colnames_substrates_combinations}
\title{Check if correct colnames are in \code{substrates_combinations}}
\usage{
.check_colnames_substrates_combinations(df, reaction = "fa_to_coa")
}
\arguments{
\item{df}{\code{data.frame}}

\item{reaction}{\code{character(1)}}
}
\description{
The helper function checks if the correct columns are in \code{df} 
depending on the \code{reaction}.
}
\details{
The function will throw an error if \code{reaction} is not of length 1 and
is not an implemented method.

\code{.check_colnames_substrates_combinations} is a helper function to 
test the integrity of \code{df}.
}
\examples{
FA <- c("FA(14:0(12Me))", "FA(16:0(14Me))", "FA(15:1(9Z)(14Me))",        
    "FA(17:0(16Me))", "FA(12:0(11Me))", "FA(13:0(12Me))", "FA(14:0(13Me))",
    "FA(15:0(14Me))", "FA(16:0(15Me))", "FA(12:0)", "FA(14:0)")
substrates <- list(FA = FA)

## create data.frame of substrates
df_substrates <- wormLipidPredictR:::.create_substrates_combinations(
    substrates = substrates, 
    constraints = "", negate = FALSE)
wormLipidPredictR:::.check_colnames_substrates_combinations(
    df = df_substrates, reaction = "fa_to_coa")
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
