% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_.check_reaction.R
\name{.check_reaction}
\alias{.check_reaction}
\title{Check if reaction is implemented}
\usage{
.check_reaction(reaction = "fa_to_coa")
}
\arguments{
\item{reaction}{\code{character(1)}}
}
\description{
The helper function checks if the \code{reaction} is in the set of 
implemented methods.

The function will throw an error if \code{reaction} is not of length 1 and
is not an implemented method.
}
\details{
\code{.check_reaction} is a helper function to test the integrity
of \code{reaction}.
}
\examples{
.check_reaction(reaction = "fa_to_coa")
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
