% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_.create_list_with_template.R
\name{.create_list_with_template}
\alias{.create_list_with_template}
\title{Create list and substitute substrates/products}
\usage{
.create_list_with_template(df_reaction, template, reaction = "RHEA:15421")
}
\arguments{
\item{df_reaction}{\code{data.frame}}

\item{template}{\code{df_template}}

\item{reaction}{\code{character(1)}}
}
\value{
list
}
\description{
Helper function for \code{create_reaction}.

The function \code{.create_list_with_template} takes the \code{template}
and replaces the \code{"reactionFormula"} by the substrates/products from
\code{df_reaction}.
}
\details{
The string replacement depend on the \code{reaction} argument.
}
\examples{
FA <- c("FA(14:0(12Me))", "FA(16:0(14Me))", "FA(15:1(9Z)(14Me))",        
    "FA(17:0(16Me))", "FA(12:0(11Me))", "FA(13:0(12Me))", "FA(14:0(13Me))",
    "FA(15:0(14Me))", "FA(16:0(15Me))", "FA(12:0)", "FA(14:0)")
substrates <- list(FA = FA)
reaction <- "RHEA:15421"

## create template
template <- LipidNetworkPredictR:::.create_template(template = NA, 
    reaction = reaction)

## create data.frame of substrates
df_substrates <- LipidNetworkPredictR:::.create_substrates_combinations(
    substrates = substrates, 
    constraints = "", negate = FALSE)
    
## add products to data.frame
df_reaction <- LipidNetworkPredictR:::.add_products(
    substrates = df_substrates, 
    reaction = reaction)
    
## make new data.frame with reaction template
LipidNetworkPredictR:::.create_list_with_template(
    df_reaction = df_reaction,
    template = template, reaction = reaction)
}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de} and 
    Thomas Naake, \email{thomasnaake@googlemail.com}
}
