% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_.create_substrates_combination.R
\name{.create_substrates_combinations}
\alias{.create_substrates_combinations}
\title{Create substrates combinations}
\usage{
.create_substrates_combinations(
  substrates = substrates,
  constraints = character(length(substrates)),
  negate = logical(length(substrates))
)
}
\arguments{
\item{substrates}{list of character vector(s)}

\item{constraints}{character vector with same length as \code{length(substrates)}}

\item{negate}{logical vector with same length as \code{length(substrates)}}
}
\value{
data.frame
}
\description{
The function returns a \code{data.frame} with all combinations from
\code{substrates}.
}
\details{
The string replacement depend on the \code{reaction} argument.

The function will internally \code{expand.grid} to create the combinations.
}
\examples{
 FA <- c("FA(14:0(12Me))", "FA(16:0(14Me))", "FA(15:1(9Z)(14Me))",        
    "FA(17:0(16Me))", "FA(12:0(11Me))", "FA(13:0(12Me))", "FA(14:0(13Me))",
    "FA(15:0(14Me))", "FA(16:0(15Me))", "FA(12:0)", "FA(14:0)")
substrates <- list(FA = FA)

## create data.frame of substrates
wormLipidPredictR:::.create_substrates_combinations(
    substrates = substrates, 
    constraints = "", negate = FALSE)
}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de} and 
    Thomas Naake, \email{thomasnaake@googlemail.com}
}
