% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_.create_template.R
\name{.create_template}
\alias{.create_template}
\title{Create template with mininum information or return a given template}
\usage{
.create_template(template = list(), reaction = "RHEA:15421")
}
\arguments{
\item{template}{\code{list}, \code{NA} or \code{NULL}}

\item{reaction}{\code{character(1)}}
}
\value{
data.frame
}
\description{
Helper function for \code{create_reaction}.

The function \code{.create_template} returns a data.frame of the template
of a given \code{reaction}. The returned data.frame is either 
the pass-through of the given \code{template} or will be created and 
filled with minimum information using the information from
\code{reaction}.
}
\details{
The function \code{.create_template} will check the \code{template} object
for the following information:

\itemize{
    \item column "reaction_name" with character entry,
    \item column "reaction_formula" with character entry,
    \item column "reaction_isReversible" with character entry, 
    \item column "reaction_geneAssociation" with character entry,
    \item column "reaction_pathway" with character entry.
}    

If \code{template} is \code{NA} or \code{NULL}, \code{template} will be set 
to an empty list.
}
\examples{
LipidNetworkPredictR:::.create_template(template = list(), reaction = "RHEA:15421")
}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
   and Thomas Naake, \email{thomasnaake@googlemail.com}
}
