% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_conversion.R
\name{find_RHEA_ids_from_SMILES}
\alias{find_RHEA_ids_from_SMILES}
\title{Find RHEA ids from SMILES}
\usage{
find_RHEA_ids_from_SMILES(
  ids,
  rhea_reaction_smiles,
  rhea_directions,
  type = c("both", "substrates", "products")
)
}
\arguments{
\item{ids}{list, entries contain the SMILES identifiers.}

\item{rhea_reaction_smiles}{data.frame containing the mappings between ChEBI 
identifiers and SMILES identifiers}

\item{rhea_directions}{data.frame containing information on the 
directionality of RHEA directions}

\item{type}{character(1), one of "both", "substrates", or "products"}
}
\value{
list
}
\description{
The function \code{find_RHEA_ids_from_SMILES} returns the RHEA ids
that are involved in reactions of the given SMILES (\code{ids}).
}
\details{
The returned result might differ depending on the \code{type} object:

\begin{itemize}
    \item if \code{type = "both"}, the function will return the RHEA 
    identifiers for the SMILES found both in the substrates or products of 
    the reactions,
    \item if \code{type = "substrates"}, the function will return the RHEA 
    identifiers for the SMILES found in the substrates of the reactions,
    \item if \code{type = "products"}, the function will return the RHEA 
    identifiers for the SMILES found in the products of the reactions.
\end{itemize}
}
\examples{
ids <- list(
    "HMDB0004947" = "CCCCCCCCCCCCC/C=C/[C@H](O)[C@H](CO)NC(=O)CCCCCCCCCCC",
    "HMDB0004949" = "CCCCCCCCCCCCC/C=C/[C@H](O)[C@H](CO)NC(=O)CCCCCCCCCCCCCCC",
    "HMDB0011763" = "CCCCCCCC/C=C\\\\CCCCCCCC(=O)N[C@H](CO)[C@H](O)CCCCCCCCCCCCCCC",
    "HMDB0004974" = NA, 
    "HMDB0011594" = NA)
    
## rhea_reaction_smiles
file <- file.path(
    path.package("LipidNetworkPredictR"), "extdata", "rhea-reaction-smiles.tsv",
    fsep = .Platform$file.sep)
rhea_reaction_smiles <- read.csv(file, sep = "\t", header = FALSE)

## rhea_directions
file <- file.path(
    path.package("LipidNetworkPredictR"), "extdata", "rhea-directions.tsv",
    fsep = .Platform$file.sep)
rhea_directions <- read.csv(file, sep = "\t", header = TRUE)

## run the function
## type = "both"
find_RHEA_ids_from_SMILES(ids = ids, 
    rhea_reaction_smiles = rhea_reaction_smiles,
    rhea_directions = rhea_directions, type = "both")
## type = "substrates"
find_RHEA_ids_from_SMILES(ids = ids, 
    rhea_reaction_smiles = rhea_reaction_smiles,
    rhea_directions = rhea_directions, type = "substrates")
## type = "products"
find_RHEA_ids_from_SMILES(ids = ids, 
    rhea_reaction_smiles = rhea_reaction_smiles,
    rhea_directions = rhea_directions, type = "products")
}
\author{
Thomas Naake
}
