% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_rhea.R
\name{getRheaIDsFromProteinID}
\alias{getRheaIDsFromProteinID}
\title{Get RHEA IDs from protein ID}
\usage{
getRheaIDsFromProteinID(uniprot_id)
}
\arguments{
\item{uniprot_id}{\code{character(1)}, UNIPROT ID, e.g. \code{"Q9H5J4"}}
}
\value{
vector
}
\description{
The function \code{getRheaIDsFromProteinID} returns associated 
RHEA IDs for a given UNIPROT ID. The function will accept IDs in the 
form of UNIPROT IDs corresponding to the 'Entry' column in Uniprot, 
e.g. "Q920L6" for fatty acid synthase from \emph{Rattus norvegicus} or
"Q9H5J4" for fatty acid synthase from \emph{Homo sapiens}.
}
\details{
The function will return an error if the protein ID is not found in 
UNIPROT. The error might be linked to internet connection problems or 
because the protein ID is incorrect.
}
\examples{
protein_id <- "Q920L6"
\dontrun{LipidNetworkPredictR:::getRheaIDsFromProteinID(protein_id)}
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
