% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_parser_lipids.R
\name{isolate_radyls}
\alias{isolate_radyls}
\title{Get all fatty acyls, alkyls and alkenyls}
\usage{
isolate_radyls(lipids)
}
\arguments{
\item{lipids}{list of vector of shorthand notation of a acyl (as string), 
e.g. \code{c("PC(18:0/20:4(7E,9E,11Z,14Z))", "PC(16:0/18:1(9Z))")}}
}
\description{
This functions isolates all fatty acyls, alkyls and alkenyls from a given 
lipid shorthand notation and returns them as vector.
}
\details{
Supported modifications are currently hydroxy groups (OH), hydroperoxy 
groups (OOH), keto groups (O), and amino groups (NH2).
}
\examples{
lipids <- c("PC(18:0/20:4(7E,9E,11Z,14Z))", "PC(16:0/18:1(9Z))")
isolate_radyls(lipids)

}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
