% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappingTable_metabolites.R
\name{mappingTable}
\alias{mappingTable}
\title{Create a mapping table to map between LipidNetworkPredictR and RHEA metabolite names}
\usage{
mappingTable()
}
\value{
data.frame
}
\description{
The function returns a \code{data.frame} containing information of 
corresponding features from \code{LipidNetworkPredictR}, the 
\code{RHEA} database and \code{ChEBI} database.
}
\details{
The mapping table is derived from the file 
`rhea-reactions.txt`.
}
\examples{
LipidNetworkPredictR:::mappingTable()
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
