% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_decouple.R
\name{obtain_ChEBI_of_feature}
\alias{obtain_ChEBI_of_feature}
\title{Obtain ChEBI ids of features}
\usage{
obtain_ChEBI_of_feature(reaction_l)
}
\arguments{
\item{reaction_l}{list as obtained from \code{create_reactions}}
}
\value{
data.frame
}
\description{
Lipid species are attributed to classes. These classes can be identified via 
the ChEBI ids in \code{reactions_l}. The function  obtains the corresponding
ChEBI ids for the features and returns a \code{data.frame} that contains 
the mappings between lipid species and lipid classes.

Mappings between the metabolites and ChEBI ids will be done via the data 
structure of \code{reaction_l}, as returned by \code{create_reactions}.
}
\details{
The function \code{obtain_ChEBI_of_feature} accepts the output of 
the \code{create_reactions} function. From \code{create_reactions}, the
ChEBI ids will be inferred.
}
\examples{
FA <- c("FA(12:0)", "FA(14:0)", "FA(16:0)")

## create data.frame with reactions and reaction order
reactions <- rbind(
    c(1, "RHEA:15421", "M_ATP + M_CoA + M_FA = M_PPi + M_AMP + M_AcylCoA", FALSE),
    c(2, "RHEA:15325", "M_Glycerol-3-P + M_AcylCoA = M_CoA + M_LPA", FALSE),
    c(3, "RHEA:19709", "M_AcylCoA + M_LPA = M_CoA + M_PA", FALSE)
)
reactions <- data.frame(order = reactions[, 1], reaction_RHEA = reactions[, 2],
    reaction_formula = reactions[, 3], directed = reactions[, 4])
reactions$order <- as.numeric(reactions$order)
reactions$directed <- as.logical(reactions$directed)

## create the list with reactions
reaction_l <- create_reactions(substrates = list(FA = FA), reactions = reactions)

## run the function
obtain_ChEBI_of_feature(reaction_l = reaction_l)
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
