% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_conversion.R
\name{obtain_substrates_products}
\alias{obtain_substrates_products}
\title{Obtain substrates and products}
\usage{
obtain_substrates_products(rhea_reaction_smiles, rhea_directions)
}
\arguments{
\item{rhea_reaction_smiles}{data.frame containing the mappings between RHEA
reaction ids and SMILES identifiers}

\item{rhea_directions}{data.frame containing information on the 
directionality of RHEA directions}
}
\value{
list
}
\description{
The function \code{obtain_substrates_products} obtains from the 
reactions stored in the \code{rhea_reaction_smiles} object the 
substrates and products.
}
\details{
The function \code{obtain_substrates_products} uses the information on the 
directionality of the reactions stored in the \code{rhea_directions}
object. This object stores in the first column reactions that have no
information on directionality, in the second column left-to-right reactions,
in the third column right-to-left reactions, and in the fourth column 
bidirectional reactions. To comply with downstream functions 
\code{obtain_substrates_products} will interpret the entries of the reactions
in \code{rhea_reaction_smiles} as in the following:

\itemize{
    \item if reaction id in column 1 of \code{rhea_directions}: 
        first entry = substrates, 
    \item if reaction id in column 2 of \code{rhea_directions}:
        first entry = substrates,
    \item if reaction id in column 3 of \code{rhea_directions}:
        first entry = products, 
    \item if reaction id in column 4 of \code{rhea_directions}:
        first entry = substrates.
}
}
\examples{
## rhea_reaction_smiles
file <- file.path(
    path.package("LipidNetworkPredictR"), "extdata", "rhea-reaction-smiles.tsv",
    fsep = .Platform$file.sep)
rhea_reaction_smiles <- read.csv(file, sep = "\t", header = FALSE)
 
## rhea_directions
file <- file.path(
    path.package("LipidNetworkPredictR"), "extdata", "rhea-directions.tsv",
    fsep = .Platform$file.sep)
rhea_directions <- read.csv(file, sep = "\t", header = TRUE)

## run the function
LipidNetworkPredictR:::obtain_substrates_products(
    rhea_reaction_smiles = rhea_reaction_smiles,
    rhea_directions = rhea_directions)
}
\author{
Thomas Naake
}
