% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_lipidmapstools.R
\name{r_GPStrGen}
\alias{r_GPStrGen}
\title{r_GPStrGen}
\usage{
r_GPStrGen(x, file_name)
}
\arguments{
\item{x}{\code{matrix} or \code{data.frame}}

\item{file_name}{\code{character(1)}}
}
\description{
r_GPStrGen
}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
