% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_conversion.R
\name{select_substrates_or_products}
\alias{select_substrates_or_products}
\title{Select substrates or products}
\usage{
select_substrates_or_products(
  reaction,
  cols,
  type = c("substrates", "products")
)
}
\arguments{
\item{reaction}{list, each entry contains a character vector of length 2}

\item{cols}{numeric of length \code{length(list)}}

\item{type}{character(1) either substrates or products}
}
\value{
list
}
\description{
The function \code{select_substrates_or_products} selects either 
substrates or products of (a) reaction(s) stored in a \code{list} object
depending on the \code{cols} argument.

Depending on the \code{type} argument the function will either return 
a list of substrates (\code{type = "substrates"}) or products
(\code{type = "products"}).
}
\details{
Each list entry stores the substrates/products or products/substrates of
a reaction. The direction of this reaction is codified in the 
\code{cols} argument. 
If \code{cols} = 1: first element is equal to substrates, 
if \code{cols} = 2: first entry is equal to substrates,
if \code{cols} = 3: first entry is equal to products, 
if \code{cols} = 4: first entry is equal to substrates.
}
\examples{
## given a rection that describes the change of (3S)-3-Hydroxy-2-butanone
## to its stereoisomer (3R)-3-Hydroxy-2-butanone
reaction <- list(c("CC(=O)[C@H](C)O",  "CC(=O)[C@H](C)O"))

## cols = 2
cols <- 2
select_substrates_or_products(reaction = reaction, cols = cols, 
    type = "substrates")
select_substrates_or_products(reaction = reaction, cols = cols, 
    type = "products")

## cols = 3
cols <- 3
select_substrates_or_products(reaction = reaction, cols = cols, 
    type = "substrates")
select_substrates_or_products(reaction = reaction, cols = cols, 
    type = "products")
}
\author{
Thomas Naake
}
