% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_conversion.R
\name{translate_ChEBI_to_SMILES}
\alias{translate_ChEBI_to_SMILES}
\title{Translate ChEBI to SMILES}
\usage{
translate_ChEBI_to_SMILES(ids, rhea_chebi_smiles)
}
\arguments{
\item{ids}{list, entries contain the}

\item{rhea_chebi_smiles}{data.frame containing the mappings between ChEBI 
identifiers and SMILES identifiers}
}
\value{
list
}
\description{
The function \code{translate_ChEBI_to_SMILES} takes ChEBI identifiers
(\code{ids}) and translate these via the dictionary \code{rhea_chebi_smiles}
to SMILES.
}
\details{
The function will return a named list. Each entry will contain the
corresponding SMILES ids. In case there are no corresponding SMILES ids
to \code{ids}, the list will contain the entry \code{NA}.
}
\examples{
ids <- list("HMDB0004947" = "72956", "HMDB0004949" = "72959", 
    "HMDB0011763" = "74100", "HMDB0004974" = "62107", 
    "HMDB0011594" = "62109")
    
## load rhea_chebi_smiles
file <- file.path(
    path.package("LipidNetworkPredictR"), "extdata", "rhea-chebi-smiles.tsv",
    fsep = .Platform$file.sep)
rhea_chebi_smiles <- read.csv(file, sep = "\t", header = FALSE)

## run the function
translate_ChEBI_to_SMILES(ids = ids, rhea_chebi_smiles = rhea_chebi_smiles)
}
\author{
Thomas Naake
}
