% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_conversion.R
\name{translate_id_to_ChEBI}
\alias{translate_id_to_ChEBI}
\title{Translate ID to ChEBI}
\usage{
translate_id_to_ChEBI(
  ids,
  id_type = colnames(annotation_set),
  annotation_set = annotation_set
)
}
\arguments{
\item{ids}{character, identifiers to be translated to ChEBI ids}

\item{id_type}{character(1), one of \code{colnames(annotation_set)}}

\item{annotation_set}{data.frame containing the mappings between different
ids}
}
\value{
list
}
\description{
The function \code{translate_id_to_ChEBI} will translate identifiers 
(\code{ids}) of type \code{id_type} to ChEBI identifiers.
}
\details{
The function will return a named list. Each entry will contain the
corresponding ChEBI ids. In case there are no corresponding ChEBI ids
to \code{ids}, the list will contain the entry \code{NA}.
}
\examples{
## define the identifiers and the id type 
ids <- c("HMDB0004947", "HMDB0004949", "HMDB0011763", "HMDB0004974", "HMDB0011594")
id_type <- "HMDB"

## define the mapping table
library(AnnotationHub)
ah <- AnnotationHub()
datasets <- query(ah, "metaboliteIDmapping")
annotation_set <- ah[["AH91792"]]

## run the function
translate_id_to_ChEBI(ids = ids, id_type = id_type,
    annotation_set = annotation_set)
}
\author{
Thomas Naake
}
