% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_reaction.R
\name{create_reaction_adjacency_matrix}
\alias{create_reaction_adjacency_matrix}
\title{Create adjacency matrix from reactions}
\usage{
create_reaction_adjacency_matrix(reaction_l)
}
\arguments{
\item{reaction_l}{list as obtained from \code{create_reactions}}
}
\value{
matrix
}
\description{
The function \code{create_reaction_adjacency_matrix} creates an adjacency
matrix connecting substrates and products that are linked by reactions.
}
\details{
The function \code{create_reaction_adjacency_matrix} accepts the output of 
the \code{create_reactions} function.
}
\examples{
FA <- c("FA(12:0)", "FA(14:0)", "FA(16:0)")

## create data.frame with reactions and reaction order
reactions <- rbind(
    c(1, "RHEA:15421", "M_atp + M_coa + M_fatacid <=> M_ppi + M_amp + M_fataccoa", FALSE),
    c(2, "RHEA:15325", "M_glyc3p + M_fataccoa <=> M_coa + M_alpa_pl", FALSE),
    c(3, "RHEA:19709", "M_fataccoa + M_alpa_pl <=> M_coa + M_pa_pl", FALSE),
    c(4, "pa_to_dg", "M_h2o + M_pa_pl <=> M_pi + M_12dag", FALSE)
)
reactions <- data.frame(order = reactions[, 1], RHEA = reactions[, 2],
    reactions = reactions[, 3], directed = reactions[, 4])
reactions$order <- as.numeric(reactions$order)
reactions$directed <- as.logical(reactions$directed)

## create the list with reactions
reaction_l <- create_reactions(substrates = list(FA = FA), reactions = reactions)

## create the adjacency matrix
create_reaction_adjacency_matrix(reaction_l)
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
