% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_reaction.R
\name{create_reactions}
\alias{create_reactions}
\title{Create reaction network for given reactions of lipid metabolism}
\usage{
create_reactions(substrates, reactions)
}
\arguments{
\item{substrates}{list containing the substrates for the reaction, 
e.g. the initial substrate for the first reaction}

\item{reactions}{data.frame containing the reaction order (in the column
\code{"order"}) and the RHEA id (in the column \code{"RHEA"}).}
}
\value{
list containing the reactions
}
\description{
The function \code{create_reactions} will create reactions following a 
given reaction order.

\code{create_reactions} will return a list containing one entry for 
each reaction type (RHEA id). Each reaction entry is again a list:
the first entry will contain a list of the products of the reaction,
the second entry will contain the filled \code{template} with the 
updated reactions.
}
\details{
The \code{data.frame} \code{reactions} has to contain the columns 
\code{order} and \code{RHEA}. It may in addition contain additional
columns such as \code{reaction} or \code{isReversible}. These columns will
be ignored by \code{create_reactions} but might be of use by the user.

The column \code{RHEA} will contain the ids that are used for 
matching the reaction type.
}
\examples{
FA <- c("FA(14:0(12Me))", "FA(16:0(14Me))", "FA(15:1(9Z)(14Me))",        
    "FA(17:0(16Me))", "FA(12:0(11Me))", "FA(13:0(12Me))", "FA(14:0(13Me))",
    "FA(15:0(14Me))", "FA(16:0(15Me))", "FA(12:0)", "FA(14:0)")

## create data.frame with reactions and reaction order
reactions <- rbind(
    c(1, "RHEA:15421", "M_atp + M_coa + M_fatacid <=> M_ppi + M_amp + M_fataccoa", FALSE),
    c(2, "RHEA:15325", "M_glyc3p + M_fataccoa <=> M_coa + M_alpa_pl", FALSE),
    c(3, "RHEA:19709", "M_fataccoa + M_alpa_pl <=> M_coa + M_pa_pl", FALSE),
    c(4, "pa_to_dg", "M_h2o + M_pa_pl <=> M_pi + M_12dag", FALSE)
)
reactions <- data.frame(order = reactions[, 1], RHEA = reactions[, 2],
    reactions = reactions[, 3], directed = reactions[, 4])
reactions$order <- as.numeric(reactions$order)
reactions$directed <- as.logical(reactions$directed)

## run the function
create_reactions(substrates = list(FA = FA), reactions = reactions)
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
