% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_via_chebi.R
\name{create_via_chebi}
\alias{create_via_chebi}
\title{create_via_chebi}
\usage{
create_via_chebi(rhea_reactions_order, rhea_reactions)
}
\arguments{
\item{rhea_reactions_order}{data.frame}

\item{rhea_reactions}{data.frame}
}
\description{
Creates ...
}
\details{
Uses ...
}
\examples{
file <- file.path(
    path.package("LipidNetworkPredictR"), "extdata", "rhea_reactions.tsv",
    fsep = .Platform$file.sep)
##rhea_reactions_order <- read.table(file, sep = "\t", header = TRUE)
##rhea_reactions_order <- rhea_reactions_order[c(1, 5, 9), ]
file <- file.path(
    path.package("LipidNetworkPredictR"), "extdata", "rhea-reactions.txt",
    fsep = .Platform$file.sep)
##rhea_reactions <- read.table(file, sep = "\t", header = FALSE, quote = "")

}
