% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_.add_products.R
\name{.add_products}
\alias{.add_products}
\title{Produce products by replacing strings of substrates}
\usage{
.add_products(substrates, reaction = "RHEA:15421")
}
\arguments{
\item{substrates}{\code{data.frame}}

\item{reaction}{\code{character(1)}}
}
\value{
data.frame
}
\description{
Helper function for \code{create_reaction}.

The function \code{.add_products} returns a data.frame of substrates and
products. The products are taken and modified from the substrates.
}
\details{
The string replacement depend on the \code{reaction} argument.

Depending on the \code{argument}, different columns (i.e. substrates)
are required and different columns (i.e. products) are returned.

The function calls internally the function
\code{stringi::stri_replace_all_fixed}.
}
\examples{
FA <- c("FA(14:0(12Me))", "FA(16:0(14Me))", "FA(15:1(9Z)(14Me))",        
    "FA(15:0(14Me))", "FA(16:0(15Me))", "FA(12:0)", "FA(14:0)")
substrates <- list(FA = FA)

## create template
template <- LipidNetworkPredictR:::.create_template(template = list(), 
    reaction = "RHEA:15421")

## create data.frame of substrates
df_substrates <- LipidNetworkPredictR:::.create_substrates_combinations(
    substrates = substrates, 
    template = template)
    
## add products to data.frame
LipidNetworkPredictR:::.add_products(substrates = df_substrates, 
    reaction = "RHEA:15421")
 
}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
   and Thomas Naake, \email{thomasnaake@googlemail.com}
}
