% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_.create_substrates_combination.R
\name{.check_colnames_substrates_combinations}
\alias{.check_colnames_substrates_combinations}
\title{Check if correct colnames are in \code{substrates_combinations}}
\usage{
.check_colnames_substrates_combinations(substrates, reaction = "RHEA:15421")
}
\arguments{
\item{substrates}{\code{data.frame} or \code{list}}

\item{reaction}{\code{character(1)}}
}
\value{
character containing the valid colnames
}
\description{
The helper function checks if the correct columns are in \code{substrates} 
depending on the \code{reaction}.
}
\details{
The function will throw an error if \code{reaction} is not of length 1 and
is not an implemented method.

The function will invisibly return the correct columns for \code{reaction}.

\code{.check_colnames_substrates_combinations} is a helper function to 
test the integrity of \code{df}.
}
\examples{
FA <- c("FA(14:0(12Me))", "FA(16:0(14Me))", "FA(15:1(9Z)(14Me))",        
    "FA(17:0(16Me))", "FA(12:0(11Me))", "FA(13:0(12Me))", "FA(14:0(13Me))",
    "FA(15:0(14Me))", "FA(16:0(15Me))", "FA(12:0)", "FA(14:0)")
substrates <- list(FA = FA)

## create data.frame of substrates
df_substrates <- LipidNetworkPredictR:::.create_substrates_combinations(
    substrates = substrates, 
    constraints = "", negate = FALSE)
LipidNetworkPredictR:::.check_colnames_substrates_combinations(
    substrates = df_substrates, reaction = "RHEA:15421")
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
